requirements.txt
Pillow==11.2.1
numpy==2.2.5
pandas==2.2.3
streamlit==1.44.1